Move outputTarget to output package
