--- generator/src/main/java/com/scottlogic/deg/generator/outputs/targets/OutputTargetValidationException.java
@@ -1,7 +1,0 @@
-package com.scottlogic.deg.generator.outputs.targets;
-
-import com.scottlogic.deg.common.ValidationException;
-
-public class OutputTargetValidationException extends ValidationException {
-    public OutputTargetValidationException(String msg) { super(msg); }
-}
