--- generator/src/main/java/com/scottlogic/deg/generator/outputs/targets/ValidatableOutput.java
@@ -1,8 +1,0 @@
-package com.scottlogic.deg.generator.outputs.targets;
-
-import java.io.IOException;
-
-/** An interface that OutputTargets implement, to reflect the fact that they're all validated in the same way */
-public interface ValidatableOutput {
-    default void validate() throws OutputTargetValidationException, IOException {}
-}
