+++ orchestrator/src/main/java/com/scottlogic/deg/orchestrator/violate/ViolateExecute.java
@@ -6,12 +6,12 @@ import com.scottlogic.deg.common.profile.Profile;
 import com.scottlogic.deg.generator.generation.DataGenerator;
 import com.scottlogic.deg.common.output.GeneratedObject;
 import com.scottlogic.deg.output.writer.DataSetWriter;
-import com.scottlogic.deg.generator.outputs.targets.SingleDatasetOutputTarget;
+import com.scottlogic.deg.output.target.SingleDatasetOutputTarget;
 import com.scottlogic.deg.orchestrator.guice.AllConfigSource;
 import com.scottlogic.deg.generator.inputs.profileviolation.ProfileViolator;
 import com.scottlogic.deg.generator.inputs.validation.ProfileValidator;
-import com.scottlogic.deg.generator.outputs.targets.OutputTargetFactory;
-import com.scottlogic.deg.generator.utils.FileUtilsImpl;
+import com.scottlogic.deg.output.target.OutputTargetFactory;
+import com.scottlogic.deg.output.FileUtils;
 import com.scottlogic.deg.generator.validators.ErrorReporter;
 import com.scottlogic.deg.orchestrator.validator.ConfigValidator;
 import com.scottlogic.deg.profile.reader.ProfileReader;
