+++ orchestrator/src/main/java/com/scottlogic/deg/orchestrator/violate/ViolateOutputValidator.java
@@ -3,8 +3,8 @@ package com.scottlogic.deg.orchestrator.violate;
 import com.google.inject.Inject;
 import com.google.inject.name.Named;
 import com.scottlogic.deg.common.profile.Profile;
-import com.scottlogic.deg.generator.outputs.targets.OutputTargetValidationException;
-import com.scottlogic.deg.generator.utils.FileUtils;
+import com.scottlogic.deg.output.target.OutputTargetValidationException;
+import com.scottlogic.deg.output.FileUtils;
 
 import java.io.IOException;
 import java.nio.file.Path;
