+++ output/build.gradle
@@ -16,6 +16,7 @@ dependencies {
 
     compile 'org.jetbrains.kotlin:kotlin-stdlib:1.3.21'
     compile 'org.apache.commons:commons-csv:1.5'
+    compile 'com.google.inject:guice:4.1.0'
 
     testCompile group: 'junit', name: 'junit', version: '4.12'
 }
